Crate buck2_subscription_proto

Modules

subscription_request

Nested message and enum types in SubscriptionRequest.

subscription_response

Nested message and enum types in SubscriptionResponse.

Structs

ActiveCommand

ActiveCommandStats

ActiveCommandsSnapshot

Disconnect

The client would like to gracefully disconnect. This will result in the command exiting.

Goodbye

/ This notification is sent by the daemon when closing the connection.

Materialized

This notification is sent by the daemon when a path that was previously passed in SubscribeToPaths is materialized.

SubscribeToActiveCommands

SubscribeToPaths

Request eager materialization of specific paths AND notifications when they get materialized. If the paths have already been materialized, a Materialized notification will be sent immediately.

SubscriptionRequest

Client-to-daemon interaction in a subscription. This is what the client writes to stdin of the subscribe command.

SubscriptionResponse

Daemon to client interaction in a subscription. This is what the client will receive via the stdout of the subscribe command.

UnsubscribeFromPaths

Undo the effects of SubscribeToPaths. In-flight notifications for the paths that are being unsubscribed from are not cancelled: the caller should expect the possibility of receiving notifications for paths after having unsubscribed from them.